backend/requirements.txt
fastapi
uvicorn
python-multipart
librosa
numpy
scipy
openai-whisper
google-generativeai
pydub
python-dotenv
soundfile
